requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
fastapi-users
bcrypt
pymongo
bson